refactor: extract HW onChain methods to a vendor agnostic protocol
